+++ charsleft_widget/tests/settings.py
@@ -5,8 +5,11 @@ SECRET_KEY = "DUMMY_SECRET_KEY"  # noqa: S105
 PROJECT_APPS = ["charsleft_widget.tests", "charsleft_widget"]
 
 INSTALLED_APPS = [
-    "django.contrib.contenttypes",
     "django.contrib.auth",
+    "django.contrib.contenttypes",
+    "django.contrib.sessions",
+    "django.contrib.messages",
+    "django.contrib.admin",
     "django.contrib.staticfiles",
     *PROJECT_APPS,
 ]
@@ -14,26 +17,12 @@ INSTALLED_APPS = [
 TEMPLATES = [
     {
         "BACKEND": "django.template.backends.django.DjangoTemplates",
-        "DIRS": [],
         "APP_DIRS": True,
-        "OPTIONS": {
-            "context_processors": [
-                "django.contrib.auth.context_processors.auth",
-                "django.template.context_processors.debug",
-                "django.template.context_processors.i18n",
-                "django.template.context_processors.media",
-                "django.template.context_processors.request",
-                "django.template.context_processors.static",
-                "django.template.context_processors.tz",
-                "django.contrib.messages.context_processors.messages",
-            ],
-        },
     },
     {
         # Mirrors the app-dirs convention a real project uses to serve
         # "charsleft_widget/textarea.jinja" from charsleft_widget/jinja2/.
         "BACKEND": "django.template.backends.jinja2.Jinja2",
-        "DIRS": [],
         "APP_DIRS": True,
     },
 ]